+++ src/index.ts
@@ -1,7 +1,7 @@
 export const version = '0.9.7';
 
 export { BaseHTMLElement } from './c-base';
-export { all, append, AppendPosition, attr, closest, elem, first, frag, next, prev, style } from './dom';
+export { all, append, AppendPosition, attr, className, closest, elem, first, frag, next, prev, style } from './dom';
 export { pull, puller, push, pusher } from './dx';
 export { addOnEvents, bindOnEvent, bindOnEvents, off, on, OnEvent, OnEventListener, OnListenerBySelector, trigger } from './event';
 export { addHubEvents, bindHubEvents, hub, Hub, HubBindings, unbindHubEvents } from './hub';